good-pass-15 Spring Data JDBC 구현
