--- src/main/java/com/goodchalk/goodpass/controller/dailypass/dto/response/DailyPassSaveFinishDto.java
@@ -1,11 +1,0 @@
-package com.goodchalk.goodpass.controller.dailypass.dto.response;
-
-import lombok.Builder;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Builder
-public class DailyPassSaveFinishDto {
-    private final String climbingGymName;
-    private final String message;
-}
