+++ src/main/java/com/goodchalk/goodpass/dailypass/controller/dto/response/DailyPassSaveFormDto.java
@@ -1,0 +1,10 @@
+package com.goodchalk.goodpass.dailypass.controller.dto.response;
+
+import lombok.Builder;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Builder
+public class DailyPassSaveFormDto {
+    private final Long climbingGymId;
+}
