+++ src/main/java/com/goodchalk/goodpass/domain/model/Contract.java
@@ -1,0 +1,5 @@
+package com.goodchalk.goodpass.domain.model;
+
+public enum Contract {
+    AGREE, DISAGREE
+}
