+++ src/main/java/com/goodchalk/goodpass/exception/GoodPassBusinessException.java
@@ -1,0 +1,6 @@
+package com.goodchalk.goodpass.exception;
+
+import lombok.RequiredArgsConstructor;
+
+public class GoodPassBusinessException extends RuntimeException {
+}
