+++ src/main/java/com/goodchalk/goodpass/exception/GoodPassSystemException.java
@@ -1,0 +1,4 @@
+package com.goodchalk.goodpass.exception;
+
+public class GoodPassSystemException extends RuntimeException{
+}
