--- src/main/java/com/goodchalk/goodpass/exception/domain/NoSuchClimbingGymException.java
@@ -1,4 +1,0 @@
-package com.goodchalk.goodpass.exception.domain;
-
-public class NoSuchClimbingGymException extends RuntimeException {
-}
