--- src/main/java/com/goodchalk/goodpass/exception/infra/NotImplementedException.java
@@ -1,4 +1,0 @@
-package com.goodchalk.goodpass.exception.infra;
-
-public class NotImplementedException extends RuntimeException{
-}
