--- src/main/java/com/goodchalk/goodpass/service/domain/Contract.java
@@ -1,5 +1,0 @@
-package com.goodchalk.goodpass.service.domain;
-
-public enum Contract {
-    AGREE, DISAGREE
-}
