+++ src/test/java/com/goodchalk/goodpass/domain/repository/ClimbingGymRepositoryTest.java
@@ -1,0 +1,29 @@
+package com.goodchalk.goodpass.domain.repository;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ClimbingGymRepositoryTest {
+    private final ClimbingGymRepository climbingGymRepository = new ClimbingGymMemoryRepository();
+
+    @Test
+    void findByClimbingGymName() {
+    }
+
+    @Test
+    void save() {
+    }
+
+    @Test
+    void findById() {
+    }
+
+    @Test
+    void deleteAll() {
+    }
+
+    @Test
+    void findAll() {
+    }
+}
