+++ src/test/java/com/goodchalk/goodpass/domain/repository/DailyPassRepositoryTest.java
@@ -1,0 +1,28 @@
+package com.goodchalk.goodpass.domain.repository;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class DailyPassRepositoryTest {
+
+    @Test
+    void findByUserName() {
+    }
+
+    @Test
+    void findByClimbingGymId() {
+    }
+
+    @Test
+    void deleteAll() {
+    }
+
+    @Test
+    void findById() {
+    }
+
+    @Test
+    void save() {
+    }
+}
